Fixed a print issue when generating listing file, if the output byte is negative number.
git-svn-id: https://edk2-buildtools.svn.sourceforge.net/svnroot/edk2-buildtools/trunk/BaseTools@1157 c2973edb-eda0-4c78-bc6a-9341b269661f

+++ Source/C/VfrCompile/VfrFormPkg.cpp
@@ -634,7 +634,7 @@ CIfrRecordInfoDB::IfrRecordOutput (
       fprintf (File, ">%08X: ", pNode->mOffset);
       if (pNode->mIfrBinBuf != NULL) {
         for (Index = 0; Index < pNode->mBinBufLen; Index++) {
-          fprintf (File, "%02X ", pNode->mIfrBinBuf[Index]);
+          fprintf (File, "%02X ", (UINT8)(pNode->mIfrBinBuf[Index]));
         }
       }
       fprintf (File, "\n");
